+++ forseti/tools/make_checksum.py
@@ -1,8 +1,8 @@
 import typing
 
-from frames.worker import ProcessJob
-from frames.tools.abstool import AbsTool
-from frames import worker
+from forseti.worker import ProcessJob
+from forseti.tools.abstool import AbsTool
+from forseti import worker
 
 class MakeChecksumBatch(AbsTool):
     name = "Make Checksum Batch"