+++ forseti/tools/zip_packages.py
@@ -4,10 +4,10 @@ import time
 
 import os
 
-from frames import worker
+from forseti import worker
 # from frames.tool import  ZipPackageJob
-from frames.tools.abstool import AbsTool
-from frames.worker import ProcessJob
+from forseti.tools.abstool import AbsTool
+from forseti.worker import ProcessJob
 
 import hathizip.process