+++ tests/test_tools.py
@@ -1,4 +1,4 @@
-from frames import tool
+from forseti import tool
 from PyQt5 import QtCore
 import pytest