+++ HISTORY.rst
@@ -6,4 +6,4 @@ Release History
 Dev
 +++
 
-* Named Project Avondale
+* Named Forseti